MCR-2449 cleanup PI-implemenation for PURLs
