+++ mycore-pi/src/main/java/org/mycore/pi/purl/MCRPURLService.java
@@ -18,20 +18,29 @@
 
 package org.mycore.pi.purl;
 
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.Date;
+import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 import java.util.function.Consumer;
+import java.util.function.Predicate;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.mycore.backend.jpa.MCREntityManagerProvider;
 import org.mycore.datamodel.metadata.MCRBase;
-import org.mycore.pi.MCRPIService;
+import org.mycore.datamodel.metadata.MCRObjectID;
+import org.mycore.pi.MCRPIJobService;
+import org.mycore.pi.backend.MCRPI;
 import org.mycore.pi.exceptions.MCRPersistentIdentifierException;
 
-public class MCRPURLService extends MCRPIService<MCRPURL> {
+public class MCRPURLService extends MCRPIJobService<MCRPURL> {
 
     private static final String TYPE = "purl";
 
-    private static final Logger LOGGER = LogManager.getLogger();
+    private static final String CONTEXT_PURL = "PURL";
+
+    private static final String CONTEXT_OBJECT = "ObjectID";
 
     private static final String PURL_SERVER_CONFIG = "Server";
 
@@ -39,16 +48,97 @@ public class MCRPURLService extends MCRPIService<MCRPURL> {
 
     private static final String PURL_PASSWORD_CONFIG = "Password";
 
-    private static final String PURL_MAINTAINER_CONFIG = "Maintainer";
-
     private static final String PURL_BASE_URL = "RegisterBaseURL";
 
-    private static final String PURL_CONTEXT_CONFIG = "RegisterContext";
+    private static final String PURL_CONTEXT_CONFIG = "RegisterURLContext";
+
+    private static final String PURL_MAINTAINER_CONFIG = "Maintainer";
 
     private static final String DEFAULT_CONTEXT_PATH = "receive/$ID";
 
     public MCRPURLService() {
         super(TYPE);
+    }
+
+    @Override
+    public void registerJob(Map<String, String> parameters) throws MCRPersistentIdentifierException {
+        MCRPURL purl = getPURLFromJob(parameters);
+        String idString = parameters.get(CONTEXT_OBJECT);
+        validateJobUserRights(MCRObjectID.getInstance(idString));
+
+        doWithPURLManager(
+            manager -> manager
+                .registerNewPURL(purl.getUrl().getPath(), buildTargetURL(idString), "302", getProperties().getOrDefault(
+                    PURL_MAINTAINER_CONFIG, "test")));
+        this.updateStartRegistrationDate(MCRObjectID.getInstance(idString), "", new Date());
+    }
+
+    protected String buildTargetURL(String objId) {
+        String baseURL = getProperties().get(PURL_BASE_URL);
+        return baseURL + getProperties().getOrDefault(PURL_CONTEXT_CONFIG, DEFAULT_CONTEXT_PATH)
+            .replaceAll("\\$[iI][dD]", objId);
+
+    }
+
+    @Override
+    protected Optional<String> getJobInformation(Map<String, String> contextParameters) {
+        return Optional.empty();
+    }
+
+    private MCRPURL getPURLFromJob(Map<String, String> parameters)
+        throws MCRPersistentIdentifierException {
+        String purlString = parameters.get(CONTEXT_PURL);
+
+        try {
+            return new MCRPURL(new URL(purlString));
+        } catch (MalformedURLException e) {
+            throw new MCRPersistentIdentifierException("Cannot parse " + purlString);
+        }
+    }
+
+    @Override
+    public void updateJob(Map<String, String> parameters) throws MCRPersistentIdentifierException {
+        String purlString = parameters.get(CONTEXT_PURL);
+        String objId = parameters.get(CONTEXT_OBJECT);
+
+        validateJobUserRights(MCRObjectID.getInstance(objId));
+        MCRPURL purl;
+
+        try {
+            purl = new MCRPURL(
+                new URL(purlString));
+        } catch (MalformedURLException e) {
+            throw new MCRPersistentIdentifierException("Could not parse purl: " + purlString, e);
+        }
+
+        doWithPURLManager((purlManager) -> {
+            if (!purlManager.isPURLTargetURLUnchanged(purl.getUrl().toString(), buildTargetURL(
+                objId))) {
+                purlManager.updateExistingPURL(purl.getUrl().getPath(), buildTargetURL(objId), "302",
+                    getProperties().getOrDefault(
+                        PURL_MAINTAINER_CONFIG, "test"));
+            }
+        });
+    }
+
+    @Override
+    public void deleteJob(Map<String, String> parameters) {
+        // deleting ist not supported
+    }
+
+    @Override
+    public MCRPI insertIdentifierToDatabase(MCRBase obj, String additional,
+        MCRPURL identifier) {
+        Date registrationStarted = null;
+        if (getRegistrationPredicate().test(obj)) {
+            registrationStarted = new Date();
+            startRegisterJob(obj, identifier);
+        }
+
+        MCRPI databaseEntry = new MCRPI(identifier.asString(), getType(), obj.getId().toString(), additional,
+            this.getServiceID(), provideRegisterDate(obj, additional), registrationStarted);
+        MCREntityManagerProvider.getCurrentEntityManager().persist(databaseEntry);
+        return databaseEntry;
     }
 
     @Override
@@ -58,25 +148,42 @@ public class MCRPURLService extends MCRPIService<MCRPURL> {
             throw new MCRPersistentIdentifierException(
                 getClass().getName() + " doesn't support additional information! (" + additional + ")");
         }
-
-        LOGGER.info("TODO: Register PURL at PURL-Server!");
-
-        doWithPURLManager(
-            manager -> {
-                if (!manager.existsPURL(purl.getUrl().getPath())) {
-                    manager.registerNewPURL(purl.getUrl().getPath(), buildTargetURL(obj), "302",
-                        getProperties().getOrDefault(PURL_MAINTAINER_CONFIG, "test"));
-                } else {
-                    manager.updateExistingPURL(purl.getUrl().getPath(), buildTargetURL(obj), "302",
-                        getProperties().getOrDefault(PURL_MAINTAINER_CONFIG, "test"));
-                }
-            });
     }
 
-    protected String buildTargetURL(MCRBase obj) {
-        String baseURL = getProperties().get(PURL_BASE_URL);
-        return baseURL + getProperties().getOrDefault(PURL_CONTEXT_CONFIG, DEFAULT_CONTEXT_PATH)
-            .replaceAll("\\$[iI][dD]", obj.getId().toString());
+    @Override
+    protected void delete(MCRPURL identifier, MCRBase obj, String additional)
+        throws MCRPersistentIdentifierException {
+        // not supported
+    }
+
+    @Override
+    protected void update(MCRPURL purl, MCRBase obj, String additional)
+        throws MCRPersistentIdentifierException {
+        if (!hasRegistrationStarted(obj.getId(), additional)) {
+            Predicate<MCRBase> registrationCondition = getRegistrationPredicate();
+            if (registrationCondition.test(obj)) {
+                this.updateStartRegistrationDate(obj.getId(), "", new Date());
+                startRegisterJob(obj, purl);
+            }
+        } else {
+            if (isRegistered(obj.getId(), "")) {
+                startUpdateJob(obj, purl);
+            }
+        }
+    }
+
+    private void startUpdateJob(MCRBase obj, MCRPURL purl) {
+        HashMap<String, String> contextParameters = new HashMap<>();
+        contextParameters.put(CONTEXT_PURL, purl.asString());
+        contextParameters.put(CONTEXT_OBJECT, obj.getId().toString());
+        this.addUpdateJob(contextParameters);
+    }
+
+    private void startRegisterJob(MCRBase obj, MCRPURL purl) {
+        HashMap<String, String> contextParameters = new HashMap<>();
+        contextParameters.put(CONTEXT_PURL, purl.asString());
+        contextParameters.put(CONTEXT_OBJECT, obj.getId().toString());
+        this.addRegisterJob(contextParameters);
     }
 
     protected void doWithPURLManager(Consumer<MCRPURLManager> action) {
@@ -94,26 +201,4 @@ public class MCRPURLService extends MCRPIService<MCRPURL> {
             manager.logout();
         }
     }
-
-    @Override
-    protected void delete(MCRPURL identifier, MCRBase obj, String additional)
-        throws MCRPersistentIdentifierException {
-        /* deletion is not supported */
-        throw new MCRPersistentIdentifierException("Delete is not supported for " + getType());
-    }
-
-    @Override
-    protected void update(MCRPURL purl, MCRBase obj, String additional)
-        throws MCRPersistentIdentifierException {
-        doWithPURLManager((purlManager) -> {
-            String targetURL = buildTargetURL(obj);
-            if (!purlManager.isPURLTargetURLUnchanged(purl.getUrl().toString(), targetURL)) {
-                purlManager.updateExistingPURL(purl.getUrl().getPath(), targetURL, "302",
-                    getProperties().getOrDefault(
-                        PURL_MAINTAINER_CONFIG, "test"));
-            }
-        });
-
-    }
-
 }
